+++ lib/ruby_speech/grxml.rb
@@ -13,6 +13,9 @@ module RubySpeech
       autoload :Token
     end
 
+    autoload :Match
+    autoload :NoMatch
+
     InvalidChildError = Class.new StandardError
 
     GRXML_NAMESPACE = 'http://www.w3.org/2001/06/grammar'
@@ -21,7 +24,7 @@ module RubySpeech
       Grammar.new(attributes).tap do |grammar|
         block_return = grammar.eval_dsl_block &block
         grammar << block_return if block_return.is_a?(String)
-      end
+      end.assert_has_matching_root_rule
     end
 
     def self.import(other)