+++ lib/ruby_speech/grxml/grammar.rb
@@ -18,37 +18,9 @@ module RubySpeech
 
       register :grammar
 
+      self.defaults = { :version => '1.0', :language => "en-US" }
+
       VALID_CHILD_TYPES = [Nokogiri::XML::Element, Nokogiri::XML::Text, Rule, Tag].freeze
-
-      ##
-      # Create a new GRXML grammar root element
-      #
-      # @param [Hash] atts Key-value pairs of options mapping to setter methods
-      #
-      # @return [Grammar] an element for use in an GRXML document
-      #
-      def self.new(atts = {}, &block)
-        new_node = super('grammar', atts)
-        new_node[:version] = '1.0'
-        new_node.namespace = GRXML_NAMESPACE
-        new_node.language ||= "en-US"
-        new_node.eval_dsl_block &block
-        new_node
-      end
-
-      ##
-      # @return [String] the base URI to which relative URLs are resolved
-      #
-      def base_uri
-        read_attr :base
-      end
-
-      ##
-      # @param [String] uri the base URI to which relative URLs are resolved
-      #
-      def base_uri=(uri)
-        write_attr 'xml:base', uri
-      end
 
       ##
       #
@@ -57,7 +29,7 @@ module RubySpeech
       # @return [String]
       #
       def mode
-        read_attr :mode
+        read_attr :mode, :to_sym
       end
 
       ##
@@ -84,17 +56,6 @@ module RubySpeech
         write_attr :root, ia
       end
 
-      def <<(arg)
-        raise InvalidChildError, "A Grammar can only accept Rule and Tag as children" unless VALID_CHILD_TYPES.include? arg.class
-        super
-      end
-
-      def to_doc
-        Nokogiri::XML::Document.new.tap do |doc|
-          doc << self
-        end
-      end
-
       ##
       #
       # @return [String]
@@ -110,11 +71,39 @@ module RubySpeech
         write_attr :'tag-format', s
       end
 
+      ##
+      # @return [Rule] The root rule node for the document
+      #
       def root_rule
         children(:rule, :id => root).first
       end
 
+      ##
+      # Checks for a root rule matching the value of the root tag
+      #
+      # @raises [InvalidChildError] if there is not a rule present in the document with the correct ID
+      #
+      # @return [Grammar] self
+      #
+      def assert_has_matching_root_rule
+        raise InvalidChildError, "A GRXML document must have a rule matching the root rule name" unless has_matching_root_rule?
+        self
+      end
+
+      ##
+      # @return [Grammar] an inlined copy of self
+      #
       def inline
+        clone.inline!
+      end
+
+      ##
+      # Replaces rulerefs in the document with a copy of the original rule.
+      # Removes all top level rules except the root rule
+      #
+      # @return self
+      #
+      def inline!
         find("//ns:ruleref", :ns => namespace_href).each do |ref|
           rule = children(:rule, :id => ref[:uri].sub(/^#/, '')).first
           ref.swap rule.nokogiri_children
@@ -126,16 +115,158 @@ module RubySpeech
         self
       end
 
-      def +(other)
-        self.class.new(:base_uri => base_uri).tap do |new_grammar|
-          (self.children + other.children).each do |child|
-            new_grammar << child
+      ##
+      # Replaces textual content of the document with token elements containing such content.
+      # This homogenises all tokens in the document to a consistent format for processing.
+      #
+      def tokenize!
+        traverse do |element|
+          next unless element.is_a? Nokogiri::XML::Text
+
+          next if self.class.import(element.parent).is_a? Token
+
+          tokens = split_tokens(element).map do |string|
+            Token.new.tap { |token| token << string }
           end
+
+          element.swap Nokogiri::XML::NodeSet.new(Nokogiri::XML::Document.new, tokens)
         end
+      end
+
+      ##
+      # Normalizes whitespace within tokens in the document according to the rules in the SRGS spec (http://www.w3.org/TR/speech-grammar/#S2.1)
+      # Leading and trailing whitespace is removed, and multiple spaces within the string are collapsed down to single spaces.
+      #
+      def normalize_whitespace
+        traverse do |element|
+          next if element === self
+
+          imported_element = self.class.import element
+          next unless imported_element.respond_to? :normalize_whitespace
+
+          imported_element.normalize_whitespace
+          element.swap imported_element
+        end
+      end
+
+      ##
+      # Checks the grammar for a match against an input string
+      #
+      # @param [String] other the input string to check for a match with the grammar
+      #
+      # @return [NoMatch, Match] depending on the result of a match attempt. If a match can be found, it will be returned with appropriate mode/confidence/utterance and interpretation attributes
+      #
+      # @example A grammar that takes a 4 digit pin terminated by hash, or the *9 escape sequence
+      #     ```ruby
+      #       grammar = RubySpeech::GRXML.draw :mode => :dtmf, :root => 'pin' do
+      #         rule :id => 'digit' do
+      #           one_of do
+      #             ('0'..'9').map { |d| item { d } }
+      #           end
+      #         end
+      #
+      #         rule :id => 'pin', :scope => 'public' do
+      #           one_of do
+      #             item do
+      #               item :repeat => '4' do
+      #                 ruleref :uri => '#digit'
+      #               end
+      #               "#"
+      #             end
+      #             item do
+      #               "\* 9"
+      #             end
+      #           end
+      #         end
+      #       end
+      #
+      #       >> subject.match '*9'
+      #       => #<RubySpeech::GRXML::Match:0x00000100ae5d98
+      #             @mode = :dtmf,
+      #             @confidence = 1,
+      #             @utterance = "*9",
+      #             @interpretation = "*9"
+      #           >
+      #       >> subject.match '1234#'
+      #       => #<RubySpeech::GRXML::Match:0x00000100b7e020
+      #             @mode = :dtmf,
+      #             @confidence = 1,
+      #             @utterance = "1234#",
+      #             @interpretation = "1234#"
+      #           >
+      #       >> subject.match '111'
+      #       => #<RubySpeech::GRXML::NoMatch:0x00000101371660>
+      #
+      #     ```
+      #
+      def match(other)
+        regex = to_regexp
+        return NoMatch.new if regex == //
+        match = regex.match other
+        return NoMatch.new unless match
+
+        Match.new :mode           => mode,
+                  :confidence     => dtmf? ? 1 : 0,
+                  :utterance      => other,
+                  :interpretation => interpret_utterance(other)
+      end
+
+      ##
+      # Converts the grammar into a regular expression for matching
+      #
+      # @return [Regexp] a regular expression which is equivalent to the grammar
+      #
+      def to_regexp
+        /^#{regexp_content.join}$/
+      end
+
+      def regexp_content
+        root_rule.children.map &:regexp_content
+      end
+
+      def dtmf?
+        mode == :dtmf
+      end
+
+      def voice?
+        mode == :voice
+      end
+
+      def <<(arg)
+        raise InvalidChildError, "A Grammar can only accept Rule and Tag as children" unless VALID_CHILD_TYPES.include? arg.class
+        super
       end
 
       def eql?(o)
         super o, :language, :base_uri, :mode, :root
+      end
+
+      def embed(other)
+        raise InvalidChildError, "Embedded grammars must have the same mode" if other.is_a?(self.class) && other.mode != mode
+        super
+      end
+
+      private
+
+      def has_matching_root_rule?
+        !root || root_rule
+      end
+
+      def interpret_utterance(utterance)
+        conversion = Hash.new { |hash, key| hash[key] = key }
+        conversion['*'] = 'star'
+        conversion['#'] = 'pound'
+
+        utterance.chars.inject [] do |array, digit|
+          array << "dtmf-#{conversion[digit]}"
+        end.join ' '
+      end
+
+      def split_tokens(element)
+        element.to_s.split(/(\".*\")/).reject(&:empty?).map do |string|
+          match = string.match /^\"(.*)\"$/
+          match ? match[1] : string.split(' ')
+        end.flatten
       end
     end # Grammar
   end # GRXML