+++ lib/ruby_speech/grxml/no_match.rb
@@ -1,0 +1,10 @@
+module RubySpeech
+  module GRXML
+    class NoMatch
+      def eql?(o)
+        o.is_a? self.class
+      end
+      alias :== :eql?
+    end
+  end
+end