+++ spec/ruby_speech/ssml/prosody_spec.rb
@@ -23,7 +23,7 @@ module RubySpeech
       describe "from a document" do
         let(:document) { '<prosody pitch="medium" contour="something" range="20Hz" rate="2" duration="10" volume="loud"/>' }
 
-        subject { Element.import parse_xml(document).root }
+        subject { Element.import document }
 
         it { should be_instance_of Prosody }
 
@@ -269,12 +269,14 @@ module RubySpeech
         end
 
         it "should accept Mark" do
-          pending
           lambda { subject << Mark.new }.should_not raise_error
         end
 
+        it "should accept P" do
+          lambda { subject << P.new }.should_not raise_error
+        end
+
         it "should accept Phoneme" do
-          pending
           lambda { subject << Phoneme.new }.should_not raise_error
         end
 
@@ -282,12 +284,15 @@ module RubySpeech
           lambda { subject << Prosody.new }.should_not raise_error
         end
 
+        it "should accept S" do
+          lambda { subject << S.new }.should_not raise_error
+        end
+
         it "should accept SayAs" do
           lambda { subject << SayAs.new(:interpret_as => :foo) }.should_not raise_error
         end
 
         it "should accept Sub" do
-          pending
           lambda { subject << Sub.new }.should_not raise_error
         end